Ensure character name has been checked before creating
While the client already ensures that the character name has to be
checked before a user can proceed, this does not need to hold true for
any other clients (e.g. bots). This now ensures that the user has
checked the name of the character they are about to create. It currently
does not handle concurrent creation of characters with the same name
however.