Use derive macro to generate (de)serialization
So far we used a custom code generator to generate the code that would
serialize and deserialize packets. This certainly worked initially when
the files were small, but became hard to manage with more and more
packets due to lack of separation options. It was not possible to re-use
 structs or enums defined in a different module which required packets
that wanted to share types to be in the same module. The XML files grew
very large making it hard to trace usages due to no IDE support. I did
make an attempt to allow for separation by being able to reference other
 modules, but as we need the full type information in a packet to
(de)serialize it, this only had very limited use.

As this already had the problem of requiring manual rebuilding on
changes, I was looking for alternatives. Serde was still not an option
for the same reason I initially chose to write a code generator.
However, a custom derive macro can provide what we would need. This
introduces two new sub-crates: serde and serde-derive. `serde` provides
the traits and types used while `serde-derive` provides the derive macro
 to generate the (de)serialization code at compile time. To avoid the
problem of requiring recursive full type information that the last
implementation needed, the traits are built to use a recursive call
pattern. Thus, only the current type needs to be known which then calls
its siblings. It currently does not generate any `::new()` functions as
shorthands, but because the source file is no longer regenerated, we can
 simply add these normally as well as any other shortcuts. Additionally,
 it does not generate the mapping between opcode and packet, but this
has been solved by providing a normal macro that generates the
respective enums.

This is not totally complete as there are some cases that could be
improved or are not handled properly. For the most part, this applies to
 nested types, which are now possible, like `Option<Vec<u32>>`. However,
 these weren't necessary yet and have an alternative by just creating a
new wrapper struct.