Improve performance with monsters and visibility
Now that we spawn the monsters automatically according to their
spawners, adding visibility really tanks the server performance. This is
 almost exclusively due to iterating through every entity for every
entity. With this change, there are several steps to improve the
situation:

1) Group entities into their region to reduce the work done by each
entity. By grouping entities, an entity would only look through at most
9 regions (the surrounding ones) and their entities, already
significantly reducing the size of entities iterated through.

2) Parallelize visibility update. Since we initially only update state
relevant to each entity individually, we can parallelize this the update
 of the entities. We still do a large initial loop to group entities,
but that is harder to parallelize.

3) Only keep entities spawned if player is nearby. The single best
improvement to the performance is reducing the amount of entities to
check for. By despawning entities in regions that are not even close to
be populated by players, we reduce both entities to sort as well as the
amount of entities we have to update.

We may want to decrease the amount even further by not applying the
visibility to monsters at all and instead manually check when we need
to, for example to check for nearby players when the monster is aggro.