Refactor packet handling in ECS
So far, we handled the packets right when we read them from the client.
This meant that all packet handling had to be done in one place, and
thus also requiring a single system with inputs required by any packet
operation. With more packets, this would cause a really large single
point of control and hard to manage.

With this refactor, this is now split into different categories which
can then be run independent of each other. Thus, only the inputs
necessary for that category of packets are required. It might add a
little overhead due to moving the packets once before actually handling
them, but I assume this to be rather small.