Store server information in database instead of config
We used the gateway server configuration as a shared central location to
 define our agent servers. This however makes it rather cumbersome to
update or add new servers. Instead, we already have a shared central
location with our database, where we can store the server
configurations. With this, servers can be added and removed at any time
without restarting the login server.