Groundwork for using skills and combat
This contains quite a few internal changes that hopefully pave the way
for the future. This includes providing a way to statically access and
thus also reference all kinds of data entries, such as items, which thus
 avoids having to look them up every time their data is necessary. It
also simplifies the structure for these data collections as they're
quite similar across the board. Lastly, the agent has been changed quite
 a bit to accommodate different kinds of actions as well as queuing
them, which is necessary as not all actions can be interrupted. This
part is still incomplete as skills themselves haven't been implemented
yet.