Send missing packets to finish client startup
Before the client can actually do most interactions, it requires a few
more packets, that we didn't send before. One is the character load
finished, which contains any items or skills that are (still) on
cooldown. The other one contains information about which client macro is
 allowed/enabled. It possibly also contains the settings, but this
hasn't been reversed yet, but is good enough for the client to accept
it.