Refactor database functions for better ergonomics
This first moves all the db functions to be associated to either the
structs they return or to unit structs that represent their goal.
Additionally, since all these are async, there was a small problem where
 functions really only needed a reference, but took an owned value to
the pool, because otherwise the called would need to extract the pool,
wrap it in an async move and then pass in a reference. Otherwise, it
would create a reference to a just-cloned copy - which wouldn't work. By
 accepting `Borrow<PgPool>` we can allow both cases; passing
`pool.clone()` directly (i.e. owned) or passing `&pool`.

This also introduces an "identifier" to servers that is used - instead
of the entity id - to identify servers by user specified ids. Because
the entity ID could end up being something completely different due to
the `ON CONFLICT`, as replace allocates a new ID.