Add commands to register user and use bcypt for passwords
This adds the option to register a user via the gateway to easily
bootstrap a new server. The only thing it still requires is a valid
database configuration. When registering a user, the password will be
hashed using bcrypt which then also needs to be compared with on login.
This means we cannot compare the password on the database side anymore
and have to do it locally. This might cause a problem of the async task
being CPU-bound due to verification, but will be ignored for now.

To include clap with derive the axum crate hate to be adjusted due to
package incompatibility between peer dependencies.