Create dockerignore for env database info
When building a docker container, we want to ignore the env file
containing the database connection info. Otherwise, sqlx will try to
connect to it when building inside the docker context, where it doesn't
have access to the local database. For those instances we already have
the offline information provided by sqlx.