Include separate rpc address in agent server entry
We previously assumed that the agent server would be accessible from the
 same address as players would use to join the server. However, in an
environment, such as docker, it would make much more sense to use the
internal network allowing the RPC port to not be exposed to the outside
world. The same applies for any kind of private networking. Thus, this
splits out a separate address that should be used for the RPC calls.