Introduce first draft of inventory change tracking
When players make in-game changes, such as moving items in their
inventory, we need to track these and update the datastore accordingly.
Since we don't have an object store where we can just dump the data
into, we need to somehow transform our inventory object into the
database representation. While we could do that by converting the
current state of the inventory to a row-like structure and replacing
what we have in the data store, that would be quite wasteful.

Instead, we keep track of what has changed and translate the changes
into database updates. This makes it easier to continuously store small
adjustments into the database as we go. If no changes to an inventory
have been made, we don't need to update anything. Additionally, we can
compare changes and evaluate if we need to update the database at all.
If an item was picked up but then removed again, no actual change has
occurred. This would otherwise require loading the inventory completely
and comparing it before making changes, increasing the load even more.