Rewrite enabled region iterator and include tests
In silkroad, the datafiles contain a large slice of bytes that define
which region is active and which isn't. The iterator we defined so far
worked in that scenario because the length of the array matched that of
the size of the region type; u16. Once you introduce tests - which
mostly likely do not use the same length - this no longer held true. As
such, it would fail and either crash or produce wrong results.

The tricky part here is that we need to check two possibly overflowing
things at the same time: the u16 of the region type and the length of
the underlying array. To help to ensure that we don't overflow one while
 we step through the list, we had to introduce a tracker if we've gone
through all regions.