Fix invalid location of monsters after movement
When trying to attack a monster that has moved, the client would move to
 a seemingly random location and attack the monster normally. This is
obviously not the desired effect. The issues was caused by the monsters
not having all the necessary components to update their position by the
movement system. We still sent the movement start packet and the client
predicted the movement of the entity correctly, but the server never
updated the position oft the monster internally, causing the player to
seemingly move to a random position. By properly applying the missing
component, which is the movement state, the movement system properly
updates the position of the monster. We will eventually need the
movement state on a monster anyway as monsters can run as well.