Extract even more general types into separate crate
We already previously created a "base" crate for game related
functionality. But this was mostly intended towards being used in the
agent. There are a few things that are even more general as they're also
 included/used in the protocol, such as the rarity of an entity. For
that, we previously had two different types; one in the protocol, one in
 the game base crate. We'd have to convert between the two all the time
which made it annoying to work with. The same goes for the region and
type id. Therefor, this provides a new base crate containing more
fundamental axioms of this domain.