Eager load navmesh related data
We currently have a sort of lazy-loaded navmesh, where we would request
the parsing and loading of a navmesh region if it hadn't already been
loaded earlier. While this does make the startup a lot faster and reduce
 memory usage, having unpredictable runtime load spikes is not desirable
 either. The amount of data we'd be loading is comparatively little and
the startup time we'd lose is also negligible. Thus, it makes much more
sense to just load everything on boot.

This adjusts the navmesh to load all enabled navmesh regions as well as
the objects they contain. While we do not use the objects yet, we might
as well load them already so that we can access them once we actually
tackle the navigation and collision. We also don't _really_ both with
error handling here, yet. We _could_ report all the failures properly,
but for now we just try our best and skip over the things we couldn't
load.

As a result, we no longer require an exclusive reference to the navmesh
in our systems. Which should, in turn, make it possible for bevy to run
more of them in parallel, if the other query parameters don't interfere.