Build state machine for entity goals
We previously had several states that were tracking more overarching
goals, such as `MoveToPickup` and `MoveToAction`. These were making
things a little complicated given they were distinct from the `Moving`
state, but essentially did the same thing. Additionally, we could not
differentiate between executing a single action and wanting to continue
doing the same action after it completed. This meant we couldn't really
implement fighting given we were missing a way to track that a player
wanted to keep auto-attacking an enemy. Thus, we needed a new system
above the current agent system, that would keep track of what the
current goal of the agent is.

When an agent has a goal, we usually need to first walk towards that
goal. Now that we have this goal separate, we can simply use the normal
agent states to implement this walking part, and it would not fit well
with other systems as it would re-use the existing `Moving` state.
Another issue we would have eventually come across is that switching
between MoveToX-states would cause the player to stop moving. This was
due to a movement finish event being sent as we were leaving a moving
state. With multiple moving states to check, this would've gotten really
 annoying to handle.

This overarching system only tracks what the player intends to do,
however, it doesn't actually know _how_ to do it. For that, it still
delegates to the agent to handle execution.

We could think about removing the MoveToEntity movement goal, given it
is now implemented through the `Mind` state and there is no way a player
 can directly request such a state. Instead, a client would do an action
 or enter the 'following' state, which both would in turn simply track
the movement themselves.