Fix pickup animation not being played
There were some issues with the pickup that caused animation to not be
played. First, we sent an `InvalidTarget` error after sending a
`Completed` on the action, which seems to have borked the client a bit.
Secondly, we didn't always send a 0x3036 packet after the pickup
completed. The latter missing being the main reason we didn't see a
pickup.