Refactor how synchronization works
So far, we used a component with several lists of changes that should be
 synchronized with the client and other players. This has several
issues. First, the placer where we change something needs to know it has
 to issue the sync as well. However, as the caller, it may not be
obvious if a change should be synchronized or not. Additionally, since
we always need an exclusive reference, we prevent several system from
being run simultaneously even though they'd otherwise not overlap.

This new system uses the built-in change tracking from bevy to collect
changes that shall be synchronized. That means the caller can simply
change a component and the component will chose to track the change or
not and later a system will be able to collect the changes that
accumulated. A caller no longer needs to know how the changes are
synchronized. We still utilized ad-hoc responses for certain
interactions, but the end goal would be to also remove the `&Client`
requirements from all game queries.

It is not without faults though. We do need to derive copy/clone for all
 packets because we can no longer generate them when we need them, but
instead have to generate them once and duplicate them later to send to
all other players. Additionally, we still need to keep some additional
context in components which we manually need to reset. In most cases,
this context helps us avoid sending too many packets, as we can
accumulate changes across a game tick. For example when synchronising
the health, if the health gets decreased twice in the same tick, we can
simply send one packet for the resulting health, instead of the two
individual changes. This is even more helpful, if a change got cancelled
 out due to another change, which means we don't have to send a packet
at all. We also need additional components to ensure we don't trigger
unnecessary changes because a component has multiple responsibilities.
This is currently especially true for the `Player` component. We can
also model several interactions as events for better inter-system
interaction as well as easier synchronisation. Lastly, while we don't do
 this yet, it would be harder to send packets only to a select amount of
 other clients. This could be useful, for example to excluded non-gm
players from seeing a player having gm invisibility enabled.