Box packets in enum
The size of an enum in rust is decided by the largest variant. As our
packet enums contain a lot of variants, some of which are very large,
it's not quite optimal. Instead, we now pack the direct packets inside a
 Box to ensure the enum stays small. However, inside the packets we
still have some variants, especially in the spawn packets, that grow
very large and should be looked at.