Implement base persistence of player data
Currently, any action happening in game is not persisted. That means, if
 a character levelled up, gained an item or simply moved, they would
restart at the same position with the same character again. That
obviously makes it impossible to progress at all. This now implements
basic persistence features to persist the character data on logout and
periodically. As of now, this covers level, experience, sp, stat points,
 health & mana, and position. Things like masteries & skills or
inventory are not yet included.