Print error instead of panicking on database update
While panicking is not _awful_ in the situation, given that the database
 has gone completely out of sync with the game server for the character,
 it might still be better - for now - to simply print the error.
Eventually, we should catch the error and try to properly resolve it,
for example, by replacing the data in the database completely with the
state in the gameserver. Until that becomes necessary, we just print it
for now.