Remove some usages of inspection and traces
I initially related a span to an execution span, where the duration of
that span is recorded. However, tracing doesn't do that, and instead
just creates an execution span for the purpose of context in logs. In
these cases we don't actually log anything, making the usage of a span
pointless. For what I imagined this would do, something like
`opentelemetry` would need to be used instead. Until then, these spans
will be removed.