Extract network related functionality into skrillax-network crates
A new family of crates was created; skrillax-network. This family
contains all the necessary functionality to set up server or client
connections using the Silkroad Online protocol. As such, the
functionality is no longer needed here and this should instead be based
on those crates.

One big change is the introduction of "protocols" which are a collection
 of packages such that the conversion and parsing of a packet to a rust
struct can be done somewhat generically. It will probably not be the
end-all-be-all of design, but it provides more options in terms of
modularity.

We unfortunately have to remove the decryptor as well, but it will end
up living in a separate repository for the future.