Use command parsing library for GM commands
We previously handled GM commands in a single big system and parsing the
 arguments manually. This was easy in the beginning but has major
annoyances: We had to work around using a `Query<&Position>` in one
command system, while we also needed `Query<&mut Position>` in another.
It's obviously possible to handle that, but the more commands would get
added, the more annoying it would become. Additionally, manually parsing
 the arguments would also become tedious.

This now uses gumroad for parsing the "subcommands" and we dispatch them
 to individual systems. While this still has some downsides (manual
registration of the systems & events), it's easier to use and doesn't
clash with queries.