Split protocol crate into gateway and agent protocol
Given we have to do a lot of derive macros here, compilation can get
quite slow. Additionally, the gateway would need to get recompiled even
if we only change stuff used by the agent in the protocol. In order to
improve the situation, this splits off two new crates: the base protocol
 and the gateway protocol.