Prefer using `try_insert` when updating/attaching components
Using `insert` can cause bevy to panic, which we obviously do not want.
It is also not easy to know if a component inserted though a command
will cause a panic or not, because it is often not clear when the
commands are being applied (i.e. where in the schedule). Given that we
expect some entities to get despawned, possibly while moving, such as
when a "chunk" gets unloaded, we should use `try_insert` instead. As we
don't care about this failing at that point, this should have no
negative impact.